Add test for basic case of user borrowing an unborrowed book.
The Library class will control interactions between users and books. This commit adds a basic test and the supporting classes (no implementation) for this functionality.

As the other search functionality seems conceptually similar to that already implemented, I decided to tackle different functionality next to flesh out the structure of the application without focusing on functionality similar to that I had already implemented as I feel doing so may lead the design in a particular direction, making things harder to refactor later.